loki: update for MultiROM compatibility
Based on https://github.com/JDCTeam/android_vendor_jdc/commit/b383c83dc386ff4ffcffb7a77bbe19cc6237d410

+++ releasetools/loki.sh
@@ -7,22 +7,31 @@
 #
 # Run loki patch on boot.img for locked bootloaders, found in loki_bootloaders
 #
+# Update for MultiROM by Mattia "AntaresOne" D'Alleva
+#
 
-egrep -q -f /system/etc/loki_bootloaders /proc/cmdline
-if [ $? -eq 0 ];then
-  echo '[*] Locked bootloader version detected.'
-  export C=/tmp/loki_tmpdir
-  mkdir -p $C
-  dd if=/dev/block/platform/msm_sdcc.1/by-name/aboot of=$C/aboot.img
-  echo '[*] Patching boot.img to with loki.'
-  /system/bin/loki_tool patch boot $C/aboot.img /tmp/boot.img $C/boot.lok || exit 1
-  echo '[*] Flashing modified boot.img to device.'
-  /system/bin/loki_tool flash boot $C/boot.lok || exit 1
-  rm -rf $C
+# Check ROM installation/upgrade:
+# if touching primary ROM flash kernel in boot partition, else skip
+export MROM=$(cat /tmp/META-INF/com/google/android/updater-script | grep "MultiROM");
+if [ "$MROM" == "" ]; then
+  echo "Installing primary ROM. Flashing kernel in boot partition..."
+  egrep -q -f /system/etc/loki_bootloaders /proc/cmdline
+  if [ $? -eq 0 ];then
+    echo '[*] Locked bootloader version detected.'
+    export C=/tmp/loki_tmpdir
+    mkdir -p $C
+    dd if=/dev/block/platform/msm_sdcc.1/by-name/aboot of=$C/aboot.img
+    echo '[*] Patching boot.img to with loki.'
+    /system/bin/loki_tool patch boot $C/aboot.img /tmp/boot.img $C/boot.lok || exit 1
+    echo '[*] Flashing modified boot.img to device.'
+    /system/bin/loki_tool flash boot $C/boot.lok || exit 1
+    rm -rf $C
+  else
+    echo '[*] Unlocked bootloader version detected.'
+    echo '[*] Flashing unmodified boot.img to device.'
+    dd if=/tmp/boot.img of=/dev/block/platform/msm_sdcc.1/by-name/boot || exit 1
+  fi
 else
-  echo '[*] Unlocked bootloader version detected.'
-  echo '[*] Flashing unmodified boot.img to device.'
-  dd if=/tmp/boot.img of=/dev/block/platform/msm_sdcc.1/by-name/boot || exit 1
+  echo "Installing ROM in MultiROM. Skipping kernel flash..."
 fi
-
 exit 0
